datagen/requirements.txt
cycler==0.10.0
github3.py==0.9.5
matplotlib==1.5.1
nltk==3.2.1
numpy==1.11.1
Pillow==3.3.0
pyparsing==2.1.8
python-dateutil==2.5.3
pytz==2016.6.1
requests==2.11.0
six==1.10.0
textblob==0.11.1
uritemplate.py==0.3.0
wordcloud==1.2.1
